feat: use Yarn v3 when creating new project
